// fe/src/lib/api/generated/model/counselingEntryMode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CounselingEntryMode = typeof CounselingEntryMode[keyof typeof CounselingEntryMode];


export const CounselingEntryMode = {
  APPOINTMENT: 'APPOINTMENT',
  WALK_IN: 'WALK_IN',
  CALLED_IN: 'CALLED_IN',
  REFERRED: 'REFERRED',
} as const;